warning
Former-commit-id: 938f89e96ad3529c36c9a721fb570f997ac585c5

+++ modules/DataLayer/Storage/GidTypes.cpp
@@ -51,7 +51,7 @@ bool CheckElementSize( ElementType et, UInt32 size )
 {
   _InitElementTypeMap( );
   const ElementTypeInfo &e = _mapEType[ et ];
-  for( int i = 0; i < e.sizes.size( ); i++ )
+  for( size_t i = 0; i < e.sizes.size( ); i++ )
     {
     if ( e.sizes[ i ] == size )
       {
